[ReactIntegration] Step 6: Tweak sandbox configuration [see details]
Webpack has been configured to have two entry points: one is the sandbox
with live reloading working nice and dandy. The other is called `application`
and will be used as the entrypoint in our rails app (with the webpack-rails
javascript / stylesheet include helpers). This second entrypoint / chunk
might not hot-reload properly (yet).