Add lodash to package.json
I'm not sure why we didn't have it in package.json - perhaps because
we're using the global from sprockets which is a gem. In any case,
I'm slowly migrating to tree-shaking lodash out of our build so we
will need it in node_modules (it was already there anyway, as a
transient dependency).