Use CDN to load vendored and rails-assets-*
The endgame is moving all of these to be managed by NPM and versioned.
At the moment, a few plugins depend on being available as globals, and
some jQuery plugins are not very UMD friendly.

Long term, we should have most of these plugins in package.json and
loaded where needed, and only load jQuery (and perhaps lodash) from
the CDN (but serving a local copy in development mode).