Workaround Timecop.freeze vs db timestamp variance
It seems like db config can timestamp precision, so I'm amending
the test to expect `consented_at` to be withih a second. It could
probably be set to be within a milisecond (since the variance I've
observed is much smaller) but I'm keeping it simple and
permissive.

Variance observed in CI:
-  expected: 2018-04-26 17:37:07.235862273 +0000
-       got: 2018-04-26 17:37:07.235862000 +0000

See https://github.com/travisjeffery/timecop/issues/146