Set sentry environment
Frontend:
If the environment variable `SENTRY_ENVIRONMENT` is not set, it will
default to `development`. When deploying, `SENTRY_ENVIRONMENT` can be
set in your CI or locally, depending on how you build your production
assets.

Backend:
The rails integration will read `SENTRY_ENVIRONMENT`, but it will
default to `Rails.env` when `SENTRY_ENVIRONMENT` is not set.