Sentry.io integration (#1151)
* Add sentry/raven to member facing pages (frontend)

* Also add sentry to ApplicationController (Rails context)

* Add sentry to campaigner facing layout

* Add sentry to login, generic, and mobile preview views

* Set raven context (sentry) in Stateless Controller

* Add redux state to sentry extra data

* Set sentry environment

Frontend:
If the environment variable `SENTRY_ENVIRONMENT` is not set, it will
default to `development`. When deploying, `SENTRY_ENVIRONMENT` can be
set in your CI or locally, depending on how you build your production
assets.

Backend:
The rails integration will read `SENTRY_ENVIRONMENT`, but it will
default to `Rails.env` when `SENTRY_ENVIRONMENT` is not set.