Release (#1189) (#1190)
* Don't run simple_format on emails body

* Update email spec

Since removing simple_format, the text is no longer wrapped in <p></p>

* [hack] check to see if $ is being clobbered somewhere. (#1185)

We're seeing a lot of errors telling us that `$` is not a function. We suspect that something's messing up with jQuery but we're not entirely sure. This is a small hack to hopefully re-assign jQuery to the `$` variable. 🤞

* Update pension emails spec

* Various bugfixes (#1187)

* import I18n from 'champaign-i18n' (alias to window.I18n in webpack)

* Deprecate legacy backbone fundraiser code.

The _fundraiser_refactor.liquid partial has been deleted and its
contents are now in _fundraiser.liquid.

All templates that referenced 'Fundraiser Refactor' must be updated
to include 'Fundraiser'.

* Add global locale data in member_facing.js

Instead of doing this in ComponentWrapper, add it immediately in member_facing.js

* Removed unused lines

* Don't attempt to assign petitionForm to jXHR response if it's a string

* Redo: "Release (#1188)"

This reverts commit 40ba347133b702c41846f69fb40c913586e63226.

* Set user context in sentry