Revert double opt-in copy for DE and EN. #gdpr
The current confirmation email for german/austrian petition signatures
is not GDPR compliant as it does not give the signatories the option to
sign yet *not* opt in to our list.

We're reverting to the previous copy, which had both links, until we
come up with a new email copy that is GDPR compliant. The copy was taken
from the commit: a0308bd

Note: French and spanish did exist in a0308bd, therefore we're using the
current copy but adding a "sign but don't opt in" link to that email.
This introduces some inconsistency, but it is temporary until we fix
this situation.